Conflicts:
	base/views/views.py
	blitz/settings.py
	blitz/staticfiles/css/.sass-cache/6586048356c03e10b95778fc89396500e673262e/style.scssc
	blitz/staticfiles/css/sass/style.scss
	blitz/staticfiles/css/stylesheets/style.css
	blitz/staticfiles/js/dashboard.js
	blitz/templates/base.html
	blitz/templates/blitz_program.html
	blitz/templates/dashboard/blitz_group_header.html
	blitz/templates/log_workout.html
	blitz/templates/trainer_dashboard.html
	spotter/templates/assign_workoutplan.html
	spotter/templates/custom_trainer_page.html
	spotter/templates/invites.html
	spotter/templates/payments.html
	spotter/templates/usage.html

+++ spotter/views.py
@@ -13,6 +13,7 @@ from django.core.files.base import ContentFile
 from django.template.loader import render_to_string
 from django.views.static import serve
 from base.emails import program_loaded, program_assigned
+from django.db.models import Q
 
 from base.models import Client, Trainer, Blitz, SalesPageContent, BlitzMember, BlitzInvitation
 from workouts.models import WorkoutSet, Lift, Workout, WorkoutPlan, WorkoutPlanWeek, WorkoutPlanDay, Exercise, ExerciseCustom, WorkoutSet, WorkoutSetCustom
@@ -48,41 +49,71 @@ def spotter_index(request):
 def spotter_payments(request):
     import balanced
 
+    trainer = request.GET.get('trainer', None)
+    month = request.GET.get('month', None)
+
+    test = True if 'test' in request.GET else None
+    charge = True if 'charge' in request.GET else None
+
+    if trainer:
+        trainer = Trainer.objects.get(pk=trainer)
+
     test = True if 'test' in request.GET else None
     charge = True if 'charge' in request.GET else None
 
     clients = []
     payments = []
-    total_cost = total_paid = float(0.0)
+    grand_total_paid = total_cost = total_paid = float(0.0)
 
     for client in Client.objects.all():
         blitz = client.get_blitz()
         if not blitz:
             continue
+        if trainer and trainer != blitz.trainer:
+            continue
         # by default ignore test/free users
         if not test and client.balanced_account_uri == '':
             continue
 
-        membership = client.blitzmember_set.all()
-
-        if membership:  # this should never be missing
-            start_date = membership[0].date_created
+        if client.date_created < blitz.begin_date:
+            start_date = blitz.begin_date
         else:
-            start_date = date.today()
+            start_date = client.date_created
+
         months = (len(list(rrule.rrule(rrule.MONTHLY, start_date, until=date.today()))))
 
+        membership = client.blitzmember_set.all()
         if not membership[0].price:   # if there was no special invitation price
-            total_cost = months * blitz.price
+            if blitz.recurring:
+                total_cost = months * blitz.price
+            else:
+                total_cost = blitz.price
         else:
-            total_cost = months * membership[0].price
+            if blitz.recurring:
+                total_cost = months * membership[0].price
+            else:
+                total_cost = membership[0].price
 
-        debits = debits = balanced.Debit.query.filter(balanced.Debit.f.meta.client_id == client.pk)
+        debits = balanced.Debit.query.filter(balanced.Debit.f.meta.client_id == client.pk)
         if debits:
             for debit in debits:
-                if 'client_id' in debit.meta:
-                    payments.append({'amount': float(debit.amount)/100, 'status': debit.status, 
-                         'created_at': debit.created_at[0:10], 'xtion': debit.transaction_number })
-                    total_paid = float(total_paid) + float(debit.amount)/100
+                if not month or int(month) == int(debit.created_at[5:7]):
+                    if 'client_id' in debit.meta:
+                        payments.append({'amount': float(debit.amount)/100, 'status': debit.status, 
+                             'created_at': debit.created_at[0:10], 'xtion': debit.transaction_number })
+                        total_paid = float(total_paid) + float(debit.amount)/100
+                        grand_total_paid += float(debit.amount)/100
+
+        refunds = balanced.Refund.query.filter(balanced.Refund.f.meta.client_id == client.pk)
+        if refunds:
+            for refund in refunds:
+                if not month or int(month) == int(refund.created_at[5:7]):
+
+                    if 'client_id' in debit.meta:
+                        payments.append({'amount': float(debit.amount)/-100, 'status': debit.status, 
+                             'created_at': debit.created_at[0:10], 'xtion': debit.transaction_number })
+                        total_paid = float(total_paid) - float(debit.amount)/100
+                        grand_total_paid -= float(debit.amount)/100
 
         clients.append({'client':client, 'blitz': blitz, 'membership': membership[0],
                         'start':start_date, 'months': months, 'payments': payments,
@@ -91,8 +122,11 @@ def spotter_payments(request):
         payments = []
         total_cost = total_paid = 0
 
+    net = float(grand_total_paid * 0.85)
+    
     return render(request, 'payments.html', 
-          {'clients' : clients, 'test' : test, 'charge' : charge })
+          {'clients' : clients, 'test' : test, 'charge' : charge, 'trainer' : trainer, 
+           'total' : grand_total_paid, 'net' : net, 'month' : month })
 
 @login_required
 def spotter_usage(request):
@@ -106,24 +140,28 @@ def spotter_usage(request):
 
     # get clients with CC on file
     paying_clients = Client.objects.filter(~Q(balanced_account_uri = ''))
-    MRR = 0
+    revenue = MRR = 0
     for payer in paying_clients:
         if payer.blitzmember_set:
             # recurring monthly charge
-            if payer.blitzmember_set.all()[0].blitz.recurring:
-                MRR += float(payer.blitzmember_set.all()[0].blitz.price)
+            if not payer.get_blitz().group and payer.blitzmember_set.all()[0].price:
+                MRR += float(payer.blitzmember_set.all()[0].price)
             # monthly charge for non-recurring blitz
             else:
-                if payer.blitzmember_set.all()[0].blitz.num_weeks() > 0:
-                    MRR += float(payer.blitzmember_set.all()[0].blitz.price / payer.blitzmember_set.all()[0].blitz.num_weeks() * 4)
+                if payer.get_blitz().num_weeks() > 0 and payer.blitzmember_set.all()[0].price:
+                    MRR += float(payer.blitzmember_set.all()[0].price / payer.blitzmember_set.all()[0].blitz.num_weeks() * 4)
+        revenue += float(payer.blitzmember_set.all()[0].price)
+
+    net = float(MRR * 0.12)
+    revenue = float(revenue * 0.12)
 
     timezone = current_tz()
     if 'days' in request.GET:
         startdate = date.today() - timedelta(days = int(request.GET.get('days')))
         days = request.GET.get('days')
     else:
-        days = 3
-        startdate = date.today() - timedelta(days = days)
+        days = 1
+        startdate = date.today() - timedelta(days = days-1)
 
     enddate = date.today() - timedelta(days=0)
     trainers = Trainer.objects.filter(date_created__range=[startdate, enddate])
@@ -139,7 +177,8 @@ def spotter_usage(request):
         usage_digest()
 
     return render(request, 'usage.html', 
-          {'days':days, 'trainers':trainers, 'login_users':login_users, 'members':members, 'MRR':MRR})
+          {'days':days, 'trainers':trainers, 'login_users':login_users, 'members':members, 
+           'revenue':revenue, 'MRR':MRR, 'net':net })
 
 @login_required
 def spotter_delete(request):
@@ -178,20 +217,94 @@ def spotter_status_trainers(request):
     trainers = Trainer.objects.all()
     return render(request, 'trainer_status.html', {'trainers' : trainers, 'errors' : None })
 
+
+@login_required
+# make copy of workoutplan
+def copy_workoutplan(request):
+    if not request.user.is_staff:
+        return redirect('home')
+
+    plan_id = request.GET.get('plan', None)
+    workoutplan = WorkoutPlan.objects.get(pk=plan_id)
+
+    if not workoutplan:
+        return redirect('home')
+    
+    wp_copy = WorkoutPlan.objects.create(trainer=workoutplan.trainer, name=workoutplan.name+' (copy)')
+
+    for week in workoutplan.workoutplanweek_set.all():
+        week_copy = WorkoutPlanWeek.objects.create(workout_plan=wp_copy, week=week.week)
+
+        for day in week.workoutplanday_set.all():
+            slug = "plan%s" % workoutplan.pk
+            if slug in day.workout.slug:
+                slug_copy = day.workout.slug.replace(slug, "plan%s" % wp_copy.pk)
+            else:
+                slug_copy = day.workout.slug + "-plan%s" % wp_copy.pk
+
+            workouts = Workout.objects.filter(slug=slug_copy)
+            if not workouts:
+                workout_copy = Workout.objects.create(display_name=day.workout.display_name, slug=slug_copy)
+            else:
+                workout_copy = workouts[0]
+
+            day_copy = WorkoutPlanDay.objects.create(workout_plan_week=week_copy, 
+                                                     workout=workout_copy,
+                                                     day_index=day.day_index,
+                                                     day_of_week=day.day_of_week)
+
+            for exercise in day.workout.exercise_set.all():            
+                exercise_copy = Exercise.objects.create(lift=exercise.lift, workout=day_copy.workout)
+                exercise_copy.sets_display = exercise.sets_display
+                exercise_copy.order = exercise.order
+
+                for set in exercise.workoutset_set.all():
+                    set_copy = WorkoutSet.objects.create(lift=set.lift, workout=day_copy.workout, exercise=exercise_copy, 
+                                                         num_reps=set.num_reps)
+
+
+    return redirect('spotter_status_trainers')
+
+
 @login_required
 def assign_workoutplan(request):
     if not request.user.is_staff:
         return redirect('home')
 
     trainers = Trainer.objects.all()
-    blitzes = Blitz.objects.all()
+
     plan_id = request.GET.get('plan', None)
     workoutplan = WorkoutPlan.objects.get(pk=plan_id)
+
+    blitz_list = []
+    blitzes = Blitz.objects.filter(trainer=workoutplan.trainer)
+    for blitz in blitzes:
+        if blitz.group:
+            blitz.title = "Group:"+blitz.title
+        if blitz.provisional:
+            blitz.title = "Provisional:"+blitz.title
+        blitz_list.append(blitz)
+
+    blitz_list.append(Blitz(title='-------------'))
+
+    blitzes = Blitz.objects.filter(~Q(trainer = workoutplan.trainer))
+    for blitz in blitzes:
+        if blitz.group:
+            blitz.title = "Group:"+blitz.title
+        if blitz.provisional:
+            blitz.title = "Provisional:"+blitz.title
+        blitz_list.append(blitz)
+
 
     if request.method == 'POST':
         form = AssignPlanForm(request.POST)
         if form.is_valid() and workoutplan:
             blitz_id = form.cleaned_data['blitz_id']
+
+            if blitz_id == 'None':
+                response = redirect('spotter_assign_workoutplan')
+                return response
+
             blitz = Blitz.objects.get(pk=blitz_id)
             blitz.workout_plan = workoutplan
             blitz.save()
@@ -203,7 +316,7 @@ def assign_workoutplan(request):
 
     form = AssignPlanForm()
     return render(request, 'assign_workoutplan.html', 
-           {'form' : form, 'workoutplan' : workoutplan, 'trainers' : trainers, 'blitzes' : blitzes })
+           {'form' : form, 'workoutplan' : workoutplan, 'trainers' : trainers, 'blitzes' : blitz_list })
 
 @login_required
 def spotter_blitz_sales_pages(request):
@@ -212,6 +325,15 @@ def spotter_blitz_sales_pages(request):
 
     pending_sales_pages = get_pending_sales_pages()  
     return render(request, 'pending_sales_pages.html', {'pending' : pending_sales_pages})
+
+@login_required
+def spotter_lifts(request):
+    print "all lifts"
+    if not request.user.is_staff:
+        return redirect('home')
+
+    lifts = Lift.objects.all() 
+    return render(request, 'all_lifts.html', {'lifts' : lifts})
 
 @login_required
 def spotter_uploads(request):
@@ -357,12 +479,41 @@ def edit_workoutplan(request):
 def workout_info(request):
 
     if request.POST.get('slug'):
-        workout = get_object_or_404(Workout, slug = request.POST.get('slug'))
-
-        if workout:
-            return JSONResponse({'num_exercises': len(workout.exercise_set.all()) })
+        workouts = Workout.objects.filter(slug=request.POST.get('slug'))
+        if workouts:
+            workout = workouts[0]
+            if workout:
+                return JSONResponse({'num_exercises': len(workout.exercise_set.all()) })
 
     return JSONResponse({'num_exercises': 0 })
+
+@csrf_exempt
+def workoutplan_rename(request):
+
+    if request.POST.get('workoutplan'):
+        workoutplan = get_object_or_404(WorkoutPlan, pk = request.POST.get('workoutplan'))
+
+        if workoutplan:
+            if request.POST.get('name'):
+                workoutplan.name = request.POST.get('name')
+                workoutplan.save()
+                print "Rename workoutplan pk=%s : %s" % (request.POST.get('workoutplan'), request.POST.get('name'))
+
+    return JSONResponse({})
+
+@csrf_exempt
+def workout_desc(request):
+
+    if request.POST.get('workout'):
+        workout = get_object_or_404(Workout, pk = request.POST.get('workout'))
+
+        if workout:
+            if request.POST.get('desc'):
+                workout.description  = request.POST.get('desc')
+                workout.save()
+                print "Description for workout pk=%s : %s" % (request.POST.get('workout'), request.POST.get('desc'))
+
+    return JSONResponse({})
 
 def new_workoutplan(request):
     flush_session_vars(request)
@@ -391,7 +542,8 @@ def view_workoutplan(request):
 
 # generate a display for workout
 def workout_display(trainer, extra):
-    return "%s %s" % (trainer.short_name, extra)
+#    return "%s %s" % (trainer.short_name, extra)
+    return extra
 
 # utility function, manages workoutplanweek/day, returns workoutplanday
 def workoutplan_day_mgr(request, workoutplan, key, workout=None, day_char=None):
@@ -611,9 +763,10 @@ def workoutplan_ajax(request):
                     exercises = Exercise.objects.filter(pk=int(request.session[session_key]))
                     print "DELETE EXERCISE REDIRECT", request.session[session_key], exercise_pk
 
-            exercise = exercises[0]
-            exercise.delete()    # delete exercise and associated workoutsets
-            print "DELETE EXERCISE", request.POST.get('key')
+            if exercises:
+                exercise = exercises[0]
+                exercise.delete()    # delete exercise and associated workoutsets
+                print "DELETE EXERCISE", request.POST.get('key')
 
     elif request.POST.get('mode') == 'add_week':
         workoutplan = get_object_or_404(WorkoutPlan, pk=request.POST.get('workoutplan'))
@@ -881,9 +1034,10 @@ def delete_plan(plan_id):
 def get_pending_sales_pages():
 
     pending_sales_pages = []
-    contents = SalesPageContent.objects.all()
+    contents = SalesPageContent.objects.all().order_by('-pk')
     for content in contents:
-        pending_sales_pages.append([content.blitz_set.all()[0], 'slug:'+content.blitz_set.all()[0].url_slug, content.name, content.trainer.name])
+        if content.blitz_set.all():
+            pending_sales_pages.append([content.blitz_set.all()[0], 'slug:'+content.blitz_set.all()[0].url_slug, content.name, content.trainer.name])
     return pending_sales_pages
